[test] 헬퍼 클래스 추가 및 테스트 코드 수정
